Order-Flow-Chain/requirements.txt
django
djangorestframework
mysqlclient
celery
redis
gunicorn
pytest-django